+++ api/controllers.py
@@ -4,9 +4,9 @@ from api import service, db
 
 bp = Blueprint('api', __name__)
 
-@bp.route('/results', methods=['POST'])
-def parse_new_results_controller():
-    added = service.parse_new_results()
+@bp.route('/draws', methods=['POST'])
+def parse_new_draws():
+    added = service.parse_new_draws()
 
     db.close()
     if added:
@@ -14,22 +14,22 @@ def parse_new_results_controller():
     else:
         return jsonify({"error": True}), 400
 
-@bp.route('/results', methods=['GET'])
-def get_results_controller():
+@bp.route('/draws', methods=['GET'])
+def get_draws():
     year = request.args.get('year')
     dates = request.args.get('dates')
 
     if dates != None:
         dates = dates.split(',')
 
-    results = service.get_results(year, dates)
+    results = service.get_draws(year, dates)
     db.close()
 
     return jsonify(results), 200
 
-@bp.route('/results/<int:contest_id>')
-def get_result_controller(contest_id):
-    contest = service.get_result(contest_id)
+@bp.route('/draws/<int:draw_id>')
+def get_draw(draw_id):
+    contest = service.get_draw(draw_id)
 
     db.close()