+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2021-04-03
+### Updated
+- All occurrences of `results` to `draws` so it's more consistent.
+- Main table `results` to `draws`
+- All endpoints from `/results` to `/draws`
+
 ## [1.0.0] - 2021-04-03
 ### Added
 - Endpoint to add new contests. This will be used as a cronjob.
@@ -36,6 +42,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial base files
 
+[2.0.0]: https://github.com/WeNeedThePoh/euromillions-api/compare/1.0.0...2.0.0
 [1.0.0]: https://github.com/WeNeedThePoh/euromillions-api/compare/0.1.0-beta...1.0.0
 [0.1.0-beta]: https://github.com/WeNeedThePoh/euromillions-api/compare/0.4.0-alpha...0.1.0-beta
 [0.4.0-alpha]: https://github.com/WeNeedThePoh/euromillions-api/compare/0.3.0-alpha...0.4.0-alpha